% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_laws.R
\name{download_laws}
\alias{download_laws}
\title{Download Current Versions of German Laws and Regulations
Download All Laws from www.gesetze-im-internet.de}
\usage{
download_laws(filetype = "xml")
}
\arguments{
\item{filetype}{Character. The type of files to be downloaded. Can be "all", "xml", "pdf" or "epub". Defaults to "xml".}
}
\description{
This function downloads all German laws and regulations from the official source www.gesetze-im-internet.de and stores them in separate folders according to file type. Supports download of XML (incl. attachments), PDF and EPUB files. Note: PDF and EPUB download are slower than XML download.
}
