% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_laws.R
\name{download_xml}
\alias{download_xml}
\title{Download XML Files from "Gesetze Im Internet"}
\usage{
download_xml(x, dir)
}
\arguments{
\item{x}{A URL to a ZIP archive on www.gesetze-im-internet.de}

\item{dir}{Target directory for download.}
}
\description{
This function transforms a link from www.gesetze-im-internet.de to use the name of the law as its filename and downloads it.
}
