% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_laws.R
\name{extract_links}
\alias{extract_links}
\title{Download Current Versions of German Laws and Regulations}
\usage{
extract_links(x, NA.omit = FALSE)
}
\description{
Download Current Versions of German Laws and Regulations
}
